requirements.txt
numpy==1.24.3
pillow==10.4.0
pygame==2.6.1
tensorflow==2.13.0